:ledger: Merge lib/file/index.js & lib/file/file.js
